requirements.txt
vvspy
pyyaml
